[mpact][compiler] add training loop to models with simple test
Note that although MPACT currently does not support autograd yet,
eventually we need to support this too. The current PR adds a very
simple training loop to the models, together with a simple neural
network that uses the training loop to learn classification of
simple sparse/dense tensors in a toy training set.